+++ eric_matthes/capitulo_5/DYS_5.5_WHILE_TRUE_4.py
@@ -19,7 +19,7 @@
 while True: # O ciclo principal controla o programa
     print("Vamos classificar a sua velocidade. Os valores aceites variam entre 0 e 200km/h")
 
-    velocidade = int(input("Diga a velocidade em km/h. Para parar, digite '999': "))
+    velocidade = float(input("Diga a velocidade em km/h. Para parar, digite '999': "))
 
     if velocidade == 999:
         print("Ordem para parar!")
@@ -28,7 +28,7 @@ while True: # O ciclo principal controla o programa
     # Validação correta: aqui SÓ avançamos com valores corretos, caso contrário continua a perguntar! NÃO DEIXA AVANÇAR! 
     while velocidade < 0 or velocidade > 200:   # Validação = while (repete até estar certo)
         print("Só são aceites valores entre 0km/h e 200km/h")
-        velocidade = int(input("Digite uma velocidade dentro deste intervalo de velocidades: "))
+        velocidade = float(input("Digite uma velocidade dentro deste intervalo de velocidades: "))
 
     # Classificação
     if velocidade <= 20:
@@ -39,4 +39,5 @@ while True: # O ciclo principal controla o programa
         print('Rápido')
     else:
         print('Muito rápido') # Podia usal um "elif", garante um resultado mais seguro!
+
