+++ eric_matthes/exercicios_cap3/EXERCÍCIO 3.7 - Pág.113.py
@@ -1,5 +1,5 @@
-n1 = int(input("Introduza um nº inteiro: " ))
-n2 = int(input("Introduza outro nº inteiro: "))
+n1 = float(input("Introduza um nº inteiro: " ))
+n2 = float(input("Introduza outro nº inteiro: "))
 soma = n1 + n2
 print(soma)
